Extract common actions to a module for reuse
A similar module exists for specs that use Poltergeist; this is useful
for those that need to use Rack::Test.
